requirements.txt
BeautifulSoup==3.2.1
Django==1.7.1
Pillow==2.8.1
PyPDF2==1.24
django-daterange-filter==1.1.1
django-model-report==0.2.1
html5lib==0.999
reportlab==3.1.44
six==1.9.0
wsgiref==0.1.2
xhtml2pdf==0.0.6
xlwt==0.7.4
